+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Delpl.hs
@@ -20,6 +20,9 @@ module Cardano.Ledger.Shelley.Rules.Delpl (
   ShelleyDelplPredFailure (..),
   ShelleyDelplEvent,
   PredicateFailure,
+
+  -- * Deprecations
+  delPlAcnt,
 )
 where
 
@@ -56,8 +59,12 @@ data DelplEnv era = DelplEnv
   { delplSlotNo :: SlotNo
   , delPlPtr :: Ptr
   , delPlPp :: PParams era
-  , delPlAcnt :: AccountState
+  , delPlAccount :: AccountState
   }
+
+delPlAcnt :: DelplEnv era -> AccountState
+delPlAcnt = delPlAccount
+{-# DEPRECATED delPlAcnt "Use `delPlAccount` instead" #-}
 
 data ShelleyDelplPredFailure era
   = PoolFailure (PredicateFailure (EraRule "POOL" era)) -- Subtransition Failures